+++ NEWS.md
@@ -3,6 +3,9 @@
 * Reworked and simplified cwd:
   * argument `thresh_terminate` removed (i.e. threshold-based event re-setting)
   * argumed `do_surplus` added
+* Clarified error message when partial year is run. simulate_snow() requires at least 365
+  days (due to snowpack spinup), but allows for partial last year.
+
 # cwd 2.0
 
 * Add `thresh_terminate_absolute` and `doy_reset` as arguments to `cwd()`